emb_server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argilla==1.10.0
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
backoff==2.2.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
cryptography==41.0.1
decorator==5.1.1
Deprecated==1.2.14
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filetype==1.2.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
nltk==3.8.1
numpy==1.23.5
olefile==0.46
openai==0.27.8
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytube==15.0.0
youtube-transcript-api==0.6.1

pytz==2023.3
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
sniffio==1.3.0
tabulate==0.9.0
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
unstructured==0.7.6
urllib3==2.0.3
uuid==1.30
validators==0.20.0
weaviate-client==3.20.1
Werkzeug==2.3.6
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0

langchain==0.0.281